Combine RemoteTestService tests
